[REFACTORING] Split IMAPServerSSLConcurrentConnectionsTest
